Only add matching fragments to the classpath
Currently all discovered fragments are added to the classpath of a
project for all os/ws combinations. But in the case of os-specific
fragments this is actually wrong and can lead to strange results as
classpath order now controls what classes are visible at compile time.

This now determines the configured environments for a project and only
add those fragments that either have no filter or match any of the given
environments.

(cherry picked from commit 6ae5ee95e78243e8cef72c88ce4fe5e7be9610d6)
